+++ fdf.h
@@ -73,47 +73,49 @@
 
 typedef struct		s_point
 {
-	int	x;
-	int	y;
-	int	z;
-	int	col;
+	long int	x;
+	long int	y;
+	long int	z;
+	long int	col;
 }					t_point;
 
 typedef struct		s_mapinfo
 {
 	char	**map;
 
-	int		lines;
-	int		width;
+	long int	lines;
+	long int	width;
 
-	int		points_num;
-	void	*mlx;
-	void	*win;
-	t_point	**points;
-	int		scale_x;
-	int		scale_y;
-	float	rot_x;
-	float	rot_y;
-	float	rot_z;
-	int		ori_x;
-	int		ori_y;
+	long int	points_num;
+	void		*mlx;
+	void		*win;
+	t_point		**points;
+	long int	scale_x;
+	long int	scale_y;
+	float		rot_x;
+	float		rot_y;
+	float		rot_z;
+	long int	ori_x;
+	long int	ori_y;
 }					t_mapinfo;
 
 typedef struct		s_line_var
 {
-	int		dx;
-	int		sx;
-	int		dy;
-	int		sy;
-	int		err;
-	int		e2;
+	long int	dx;
+	long int	sx;
+	long int	dy;
+	long int	sy;
+	long int	err;
+	long int	e2;
 }					t_line_var;
 
-t_point				new_point(int x, int y, int z);
-void				translate_point(int x, int y, int z, t_mapinfo *map);
+t_point				new_point(long int x, long int y, long int z);
+void				translate_point(long int x, long int y, long int 
+						z, t_mapinfo *map);
 void				translate_rot(float angle, t_mapinfo *map,
 						void (f)(float, t_mapinfo *));
-void				scale_points(t_mapinfo *map, int scale_x, int scale_y);
+void				scale_points(t_mapinfo *map, long int scale_x, 
+						long int scale_y);
 void				rotate_x(float angle, t_mapinfo *map);
 void				rotate_y(float angle, t_mapinfo *map);
 void				rotate_z(float angle, t_mapinfo *map);
@@ -122,7 +124,8 @@ void				ft_map(t_mapinfo *map, t_point ***cloud);
 void				ft_init_map(t_mapinfo *mapinfo);
 int					ft_argcheck(int argc, char **argv);
 void				draw_line(t_point v1, t_point v2, t_mapinfo *m);
-void				ft_movemap(t_mapinfo *map, int offset_x, int offset_y);
+void				ft_movemap(t_mapinfo *map, long int offset_x, 
+						long int offset_y);
 void				ft_centremap(t_mapinfo *map);
 void				ft_draw_map(t_mapinfo map);
 void				ft_draw_instruct(t_mapinfo map);
